Initial commit
Initial commit, adding the basic setup I was using in the online editor.
This includes:
Character -> a class to manage a split flap character, that will track
what it is currently displaying and what is the target value.
Display -> has a list of characters, and communicates the message that
wants to be displayed to each letter.
MessageRequester -> looks up in the feeds for new messages, and with
some time constraints, informs to the display when it wants to show a
message.
MotorController -> will take care of the low leve; communication with a
motor